一括設定対応

+++ JisuCounter/JisuCounterData/MS_KYOUKA.cs
@@ -15,7 +15,7 @@ namespace JisuCounterData
 
         public int MS_KYOUKA_ID { get; set; }
         public string KYOUKA_NAME { get; set; }
-        public double KYOUKA_RATIO { get; set; }
+        public float KYOUKA_RATIO { get; set; }
         public string COLOR { get; set; }
 
     }
